:truck: Move SavePairButton to ComparisonPanel.jsx
