+++ src/components/ComparisonPanel.jsx
@@ -1,7 +1,11 @@
-import { useCallback, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
+import { usePair } from '../hooks/PairContext';
+import { usePairStore } from '../hooks/usePairStore';
 import { AlphaNumComparisonSection } from './AlphaNumComparisonSection';
 import { TextualExampleComparisonSection } from './TextualExampleComparisonSection';
 import { CodeSection } from './CodeSection';
+import { BookmarkIcon as OutlineBookmarkIcon } from '@heroicons/react/24/outline';
+import { BookmarkIcon as SolidBookmarkIcon } from '@heroicons/react/24/solid';
 
 import alphaNumericalSectionIcon from '../assets/img/alpha-numerical-section-icon.svg';
 import textualSectionIcon from '../assets/img/textual-section-icon.svg';
@@ -9,6 +13,7 @@ import exportSectionIcon from '../assets/img/export-section-icon.svg';
 
 import styles from '/src/assets/styles/comparison-panel.module.css';
 import inputStyles from '/src/assets/styles/input.module.css';
+import iconStyles from '/src/assets/styles/icon.module.css';
 
 export function ComparisonPanel() {
 	const [currentSection, setCurrentSection] = useState('alphanum');
@@ -33,10 +38,13 @@ export function ComparisonPanel() {
 				/>
 				<CodeSection isCurrentSection={currentSection === 'code'} />
 			</div>
+			<>
 			<ComparisonSectionSelector
 				currentSection={currentSection}
 				onChange={handleComparisonSectionChange}
 			/>
+				<SavePairButton />
+			</>
 		</>
 	);
 }
@@ -113,5 +121,49 @@ function ComparisonSectionSelector({ currentSection, onChange }) {
 				/>
 			</label>
 		</fieldset>
+	);
+}
+
+/**
+ * Renders a Save button to save the current Pair of fonts.
+ */
+function SavePairButton() {
+	const pair = usePair();
+	const { pairs, add, remove, includes } = usePairStore();
+	const isPairInPairStore = includes(pair);
+	const [isSaved, setIsSaved] = useState(isPairInPairStore);
+
+	useEffect(() => {
+		setIsSaved(includes(pair));
+	}, [pair, pairs]);
+
+	function handleClick(e) {
+		try {
+			const pair = JSON.parse(e.target.dataset.pair);
+
+			if (isSaved) {
+				remove(pair);
+				setIsSaved(includes(pair));
+			} else {
+				add(pair);
+				setIsSaved(includes(pair));
+			}
+		} catch (error) {
+			console.error(error);
+		}
+	}
+
+	return (
+		<button
+			className={[inputStyles.buttonIcon, styles.savePairButton].join(' ')}
+			data-pair={JSON.stringify(pair || '')}
+			onClick={handleClick}
+		>
+			{isSaved ? (
+				<SolidBookmarkIcon className={iconStyles.smallIcon} />
+			) : (
+				<OutlineBookmarkIcon className={iconStyles.smallIcon} />
+			)}
+		</button>
 	);
 }
